+++ Libplanet.RocksDBStore.Tests/RocksDBStoreTest.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using System.Linq;
 using Libplanet.Blockchain;
 using Libplanet.Blockchain.Policies;
 using Libplanet.Store;
@@ -7,6 +8,7 @@ using Libplanet.Tests.Blockchain;
 using Libplanet.Tests.Common.Action;
 using Libplanet.Tests.Store;
 using Libplanet.Tests.Store.Trie;
+using Libplanet.Tx;
 using Xunit;
 using Xunit.Abstractions;
 
@@ -65,6 +67,70 @@ namespace Libplanet.RocksDBStore.Tests
             }
             finally
             {
+                Directory.Delete(path, true);
+            }
+        }
+
+        [SkippableFact]
+        public void ParallelGetBlock()
+        {
+            var path = Path.Combine(Path.GetTempPath(), $"rocksdb_test_{Guid.NewGuid()}");
+            var store = new RocksDBStore(path);
+            try
+            {
+                Guid cid = Guid.NewGuid();
+                store.AppendIndex(cid, Fx.Block1.Hash);
+                store.AppendIndex(cid, Fx.Block2.Hash);
+                store.AppendIndex(cid, Fx.Block3.Hash);
+
+                store.PutBlock(Fx.Block1);
+                store.PutBlock(Fx.Block2);
+                store.PutBlock(Fx.Block3);
+
+                store.Dispose();
+                store = new RocksDBStore(path);
+
+                Enumerable.Range(0, 3).AsParallel().ForAll(i =>
+                {
+                    var bHash = store.IndexBlockHash(cid, i).Value;
+                    var block = store.GetBlock<DumbAction>(bHash);
+                    Assert.NotNull(block);
+                });
+            }
+            finally
+            {
+                store.Dispose();
+                Directory.Delete(path, true);
+            }
+        }
+
+        [SkippableFact]
+        public void ParallelGetTransaction()
+        {
+            var path = Path.Combine(Path.GetTempPath(), $"rocksdb_test_{Guid.NewGuid()}");
+            var store = new RocksDBStore(path);
+            Transaction<DumbAction>[] txs = new[]
+            {
+                Fx.Transaction1,
+                Fx.Transaction2,
+                Fx.Transaction3,
+            };
+            try
+            {
+                store.PutTransaction(Fx.Transaction1);
+                store.PutTransaction(Fx.Transaction2);
+                store.PutTransaction(Fx.Transaction3);
+                store.Dispose();
+                store = new RocksDBStore(path);
+
+                Enumerable.Range(0, 3).AsParallel().ForAll(i =>
+                {
+                    Assert.NotNull(store.GetTransaction<DumbAction>(txs[i].Id));
+                });
+            }
+            finally
+            {
+                store.Dispose();
                 Directory.Delete(path, true);
             }
         }