+++ CHANGES.md
@@ -248,6 +248,9 @@ To be released.
     [[#1278]]
  -  Fixed a bug where `DefaultStore.ForkBlockIndexes()` hadn't copied genesis
     block.  [[#1325]]
+ -  (Libplanet.RocksDBStore) Fixed a bug where `RocksDBStore.GetBlock<T>()`
+    and `RocksDBStore.GetTransaction<T>()` handn't returned expected values
+    in multithreading environment.  [[#1339], [#1342]]
 
 ### CLI tools
 
@@ -292,6 +295,8 @@ To be released.
 [#1305]: https://github.com/planetarium/libplanet/pull/1305
 [#1325]: https://github.com/planetarium/libplanet/pull/1325
 [#1328]: https://github.com/planetarium/libplanet/pull/1328
+[#1339]: https://github.com/planetarium/libplanet/issues/1339
+[#1342]: https://github.com/planetarium/libplanet/pull/1342
 
 
 Version 0.11.1